requirements.txt
PyQt5
pymupdf
pdf2image
numpy
scikit-image
Pillow
python-docx